Just use a HTML image instead
We don’t need the base styles of our Image component. All the aspect ratio stuff in there is irrelevant, because we reset it here. The phantom whitespace thing is not an issue, because this image is a child of a wrapper with an aspect ratio. And the alt text of this image is an empty string.